feat(prophecies/units): Add modifiers in maths request
